Move function to helper to avoid duplication
